Keep original apex files even after runnint the script several times

+++ aosp_post_build_injector.py
@@ -358,7 +358,8 @@ def handle_apex_modules(file_path, aosp_path, lunch_target):
                 error_message = f"Error signing APEX file: {file_path}|{signing_key}|{signing_key_path}|{log_message}"
             else:
                 logging.info(f"APEX file signed: {file_path} with key: {signing_key}")
-                shutil.move(file_path, f"{file_path}.original_apex")
+                if not os.path.exists(f"{file_path}.original_apex"):
+                    shutil.move(file_path, f"{file_path}.original_apex")
                 shutil.move(apex_out_file, file_path)
     return error_message
